fix: move interface to types & format
