+++ src/types.d.ts
@@ -346,3 +346,30 @@ export interface ChainConfig {
   blockExplorer?: string;
   icon?: string;
 }
+
+// hook types
+
+export interface ProposalCalls {
+  from: string;
+  to: string;
+  data: string;
+  value: BigNumber;
+  recommendedCallUsed?: RecommendedCallUsed | undefined;
+  callParameters?: unknown | undefined;
+  encodedFunctionName?: string | undefined;
+}
+
+export interface RecommendedCallUsed {
+  asset: string;
+  from: string;
+  to: string;
+  toName: string;
+  functionName: string;
+  params: {
+    type: string;
+    name: string;
+    defaultValue: string;
+    decimals?: number;
+  }[];
+  decodeText: string;
+}
